move UI components into their own folder
