+++ src/renderer/components/content-types/contact/ContactEditor.tsx
@@ -14,24 +14,24 @@ import { ContentProps } from '../../Content'
 import { ContactDoc } from '.'
 import { FileDoc } from '../files'
 
-import ColorPicker from '../../ColorPicker'
+import ColorPicker from '../../ui/ColorPicker'
 import { useDocument } from '../../../Hooks'
-import Heading from '../../Heading'
-import SecondaryText from '../../SecondaryText'
+import Heading from '../../ui/Heading'
+import SecondaryText from '../../ui/SecondaryText'
 
 import { CurrentDeviceContext } from '../workspace/Device'
 import { importFileList } from '../../../ImportData'
 import ConnectionStatusBadge from './ConnectionStatusBadge'
 import { useConnectionStatus } from '../../../PresenceHooks'
-import Badge from '../../Badge'
-import CenteredStack from '../../CenteredStack'
+import Badge from '../../ui/Badge'
+import CenteredStack from '../../ui/CenteredStack'
 import { without } from '../../../Misc'
 import ContactEditorDevice, { OnRemoveDevice } from './ContactEditorDevice'
-import ListMenuSection from '../../ListMenuSection'
-import ListMenuItem from '../../ListMenuItem'
+import ListMenuSection from '../../ui/ListMenuSection'
+import ListMenuItem from '../../ui/ListMenuItem'
 import TitleEditor from '../../TitleEditor'
-import ListItem from '../../ListItem'
-import ListMenu from '../../ListMenu'
+import ListItem from '../../ui/ListItem'
+import ListMenu from '../../ui/ListMenu'
 import { USER_COLORS } from './Constants'
 import SharesSection from './SharesSection'
 import './ContactEditor.css'
