+++ src/renderer/components/ui/Badge.tsx
