+++ src/renderer/components/ui/CenteredStack.css
