+++ src/renderer/components/ui/CenteredStack.tsx
