+++ src/renderer/components/ui/ColorPicker.css
