+++ src/renderer/components/ui/ColorPicker.tsx
