+++ src/renderer/components/ui/ListMenuSection.css
@@ -1,6 +1,3 @@
-.ListMenuSection {
-}
-
 .ListMenuSection-title {
   display: flex;
   align-items: center;
